+++ meta/data.go
@@ -344,7 +344,7 @@ func (data *Data) CreateShardGroup(database, policy string, timestamp time.Time)
 		si := &sgi.Shards[i]
 		for j := 0; j < replicaN; j++ {
 			nodeID := data.Nodes[nodeIndex%len(data.Nodes)].ID
-			si.OwnerIDs = append(si.OwnerIDs, nodeID)
+			si.Owners = append(si.Owners, ShardOwner{NodeID: nodeID})
 			nodeIndex++
 		}
 	}
@@ -942,14 +942,14 @@ func (sgi *ShardGroupInfo) unmarshal(pb *internal.ShardGroupInfo) {
 
 // ShardInfo represents metadata about a shard.
 type ShardInfo struct {
-	ID       uint64
-	OwnerIDs []uint64
+	ID     uint64
+	Owners []ShardOwner
 }
 
 // OwnedBy returns whether the shard's owner IDs includes nodeID.
 func (si ShardInfo) OwnedBy(nodeID uint64) bool {
-	for _, id := range si.OwnerIDs {
-		if id == nodeID {
+	for _, so := range si.Owners {
+		if so.NodeID == nodeID {
 			return true
 		}
 	}
@@ -960,9 +960,11 @@ func (si ShardInfo) OwnedBy(nodeID uint64) bool {
 func (si ShardInfo) clone() ShardInfo {
 	other := si
 
-	if si.OwnerIDs != nil {
-		other.OwnerIDs = make([]uint64, len(si.OwnerIDs))
-		copy(other.OwnerIDs, si.OwnerIDs)
+	if si.Owners != nil {
+		other.Owners = make([]ShardOwner, len(si.Owners))
+		for i := range si.Owners {
+			other.Owners[i] = si.Owners[i].clone()
+		}
 	}
 
 	return other
@@ -974,17 +976,64 @@ func (si ShardInfo) marshal() *internal.ShardInfo {
 		ID: proto.Uint64(si.ID),
 	}
 
-	pb.OwnerIDs = make([]uint64, len(si.OwnerIDs))
-	copy(pb.OwnerIDs, si.OwnerIDs)
+	pb.Owners = make([]*internal.ShardOwner, len(si.Owners))
+	for i := range si.Owners {
+		pb.Owners[i] = si.Owners[i].marshal()
+	}
 
 	return pb
+}
+
+// UnmarshalBinary decodes the object from a binary format.
+func (si *ShardInfo) UnmarshalBinary(buf []byte) error {
+	var pb internal.ShardInfo
+	if err := proto.Unmarshal(buf, &pb); err != nil {
+		return err
+	}
+	si.unmarshal(&pb)
+	return nil
 }
 
 // unmarshal deserializes from a protobuf representation.
 func (si *ShardInfo) unmarshal(pb *internal.ShardInfo) {
 	si.ID = pb.GetID()
-	si.OwnerIDs = make([]uint64, len(pb.GetOwnerIDs()))
-	copy(si.OwnerIDs, pb.GetOwnerIDs())
+
+	// If deprecated "OwnerIDs" exists then convert it to "Owners" format.
+	if len(pb.GetOwnerIDs()) > 0 {
+		si.Owners = make([]ShardOwner, len(pb.GetOwnerIDs()))
+		for i, x := range pb.GetOwnerIDs() {
+			si.Owners[i].unmarshal(&internal.ShardOwner{
+				NodeID: proto.Uint64(x),
+			})
+		}
+	} else if len(pb.GetOwners()) > 0 {
+		si.Owners = make([]ShardOwner, len(pb.GetOwners()))
+		for i, x := range pb.GetOwners() {
+			si.Owners[i].unmarshal(x)
+		}
+	}
+}
+
+// ShardOwner represents a node that owns a shard.
+type ShardOwner struct {
+	NodeID uint64
+}
+
+// clone returns a deep copy of so.
+func (so ShardOwner) clone() ShardOwner {
+	return so
+}
+
+// marshal serializes to a protobuf representation.
+func (so ShardOwner) marshal() *internal.ShardOwner {
+	return &internal.ShardOwner{
+		NodeID: proto.Uint64(so.NodeID),
+	}
+}
+
+// unmarshal deserializes from a protobuf representation.
+func (so *ShardOwner) unmarshal(pb *internal.ShardOwner) {
+	so.NodeID = pb.GetNodeID()
 }
 
 // ContinuousQueryInfo represents metadata about a continuous query.