+++ CHANGELOG.md
@@ -16,6 +16,7 @@ With this release InfluxDB is moving to Go 1.5.
 - [#3868](https://github.com/influxdb/influxdb/pull/3868): Add shell option to start the daemon on CentOS. Thanks @SwannCroiset.
 - [#3886](https://github.com/influxdb/influxdb/pull/3886): Prevent write timeouts due to lock contention in WAL
 - [#3574](https://github.com/influxdb/influxdb/issues/3574): Querying data node causes panic
+- [#3913](https://github.com/influxdb/influxdb/issues/3913): Convert meta shard owners to objects
 
 ## v0.9.3 [2015-08-26]